修复16进制转换和函数多维数组参数的问题

+++ src/compiler/Util.java
@@ -28,12 +28,12 @@ public class Util {
         }
         if(num.startsWith("0x") || num.startsWith("0X"))
         {
-            return Integer.parseInt(num.toLowerCase(Locale.ROOT).replace("0x", ""), 16);
+            return (int)Long.parseLong(num.toLowerCase(Locale.ROOT).replace("0x", ""), 16);
         }else if(num.startsWith("0") && num.length()>1)
         {
-            return Integer.parseInt(num.toLowerCase(Locale.ROOT).substring(1),8);
+            return (int)Long.parseLong(num.toLowerCase(Locale.ROOT).substring(1),8);
         }
-        int i = Integer.parseInt(num);
+        int i = (int)Long.parseLong(num);
         return neg?-i:i;
     }
 }
